Use different endpoint for cran_package_history()
Hopefully it helps with #117.

I have no clue why CloudFlare returns 400 for the
repeated requests for the other endpoint,
and with their free tier I can't seem to be able
to dig deeper in the logs.

Nevertheless this entpoint seems to work better.

We can't easily turn off Cloudflare right now, we
would need to update the crandb server for that.